// projekat2023FEB/app/src/app/models/user.ts
export class User {
    firstname: string;
    lastname: string;
    username: string;
    password: string;
    phone: string;
    email: string;
    picture: string;

    isOrganizer: string;
    organizationName: string;
    address: string;
    IDorganization: string;

    status: string;

    isAdmin: boolean;
}